requirements.txt
SpeechRecognition==3.11.0
pyttsx3==2.90
google-genai>=1.0.0
PyAudio
python-dotenv